requirements.txt
asgiref==3.5.2
certifi==2022.6.15.2
charset-normalizer==2.1.1
Django==4.1.1
djangorestframework==3.13.1
docopt==0.6.2
idna==3.4
pipreqs==0.4.11
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
yarg==0.1.9